Separated downloader class
